proyecto_final_backend/requirements.txt
django-request
responses
asgiref==3.2.7
beautifulsoup4==4.9.1
Django==3.0.7
django-cors-headers==3.4.0
django-fsm==2.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.6.1
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
selenium==3.141.0
urllib3==1.25.10
pyunitreport~=0.1.4